validate_role

+++ web/app/views.py
@@ -127,6 +127,7 @@ def index():
             lastname = user.lastname
             user_data_ = {
                 'id' : user.id,
+                'role' : user.role,
                 'name': firstname + " " + lastname,
                 'email': usr_email
             }
@@ -156,93 +157,99 @@ def index():
 
 @app.route('/form' , methods=('GET', 'POST'))
 def form():
-    if request.method == 'POST':
-        doc_data = request.files.get('doc_data')
-        #app.logger.debug("doc data :",doc_data)
-        validated = True
-        validated_dict = dict()
-                # Read the contents of the uploaded file as bytes
-        #ref_num = order_info.query.order_by(desc(order_info.id)).first().ref_num
+    access_token = session['access_token']
+    user_data = get_user_data(access_token)
+    email = user_data.get('cmuitaccount')
+    user = User.query.filter_by(email=email).first()
+    if user.role == True:
+        if request.method == 'POST':
+            doc_data = request.files.get('doc_data')
+            #app.logger.debug("doc data :",doc_data)
+            validated = True
+            validated_dict = dict()
+                    # Read the contents of the uploaded file as bytes
+            #ref_num = order_info.query.order_by(desc(order_info.id)).first().ref_num
 
-        valid_keys = ['subject','ref_num', 'doc_date' ,'ref_year','user_id']
+            valid_keys = ['subject','ref_num', 'doc_date' ,'ref_year','user_id']
 
-        # Access the uploaded file using request.files
-        name_list =  request.form.get('name_list')
-        name_list = name_list.split(",")
-        name_list =  [i for i in name_list if i != ""]
-        #app.logger.debug("name_list = ",name_list)
-        # validate the input
-        for key in request.form:
-            if key not in valid_keys:
-                continue
-            value = request.form[key].strip()
-            if not value or value == 'undefined':
-                validated = False
-                break
-            app.logger.debug(value)
-            validated_dict[key] = value
+            # Access the uploaded file using request.files
+            name_list =  request.form.get('name_list')
+            name_list = name_list.split(",")
+            name_list =  [i for i in name_list if i != ""]
+            #app.logger.debug("name_list = ",name_list)
+            # validate the input
+            for key in request.form:
+                if key not in valid_keys:
+                    continue
+                value = request.form[key].strip()
+                if not value or value == 'undefined':
+                    validated = False
+                    break
+                app.logger.debug(value)
+                validated_dict[key] = value
 
-        if validated:
-            order = order_info.query.filter(and_(order_info.ref_num == int(validated_dict['ref_num']) , \
-                                             order_info.ref_year == int(validated_dict['ref_year']))).first()
-            app.logger.debug("order :",order)
-            if order == None:
-                empty_order = order_info.query.filter(order_info.subject == None).first()
-                #app.logger.debug("empty_order", empty_order)
-                if empty_order != None:
-                    app.logger.debug(empty_order)
-                    db.session.delete(empty_order)
+            if validated:
+                order = order_info.query.filter(and_(order_info.ref_num == int(validated_dict['ref_num']) , \
+                                                order_info.ref_year == int(validated_dict['ref_year']))).first()
+                app.logger.debug("order :",order)
+                if order == None:
+                    empty_order = order_info.query.filter(order_info.subject == None).first()
+                    #app.logger.debug("empty_order", empty_order)
+                    if empty_order != None:
+                        app.logger.debug(empty_order)
+                        db.session.delete(empty_order)
+                        db.session.commit()
+                    order_entry = order_info(
+                    subject=validated_dict['subject'],
+                    doc_date=validated_dict['doc_date'],
+                    ref_num= validated_dict['ref_num'],
+                    ref_year=validated_dict['ref_year'],
+                    ref_name=name_list,
+                    user_id=validated_dict['user_id']
+                    )
+                    db.session.add(order_entry)
                     db.session.commit()
-                order_entry = order_info(
-                subject=validated_dict['subject'],
-                doc_date=validated_dict['doc_date'],
-                ref_num= validated_dict['ref_num'],
-                ref_year=validated_dict['ref_year'],
-                ref_name=name_list,
-                user_id=validated_dict['user_id']
-                )
-                db.session.add(order_entry)
-                db.session.commit()
-                if doc_data != None :
-                    doc_content = doc_data.read()
-                    doc_entry = doc_info(
-                    order_refnum = order_entry.ref_num,
-                    order_refyear = order_entry.ref_year,
-                    filename= str(validated_dict['ref_num'])+"/"+str(validated_dict['ref_year']),
-                    doc_data=doc_content
-                )
-                    db.session.add(doc_entry)   
-                
-            else:
-                app.logger.debug("update")
-
-                order_entry = order.update(
-                subject=validated_dict['subject'],
-                doc_date=validated_dict['doc_date'],
-                ref_name=name_list,
-                user_id=validated_dict['user_id']
-                )
-                if doc_data != None :
-                    doc_content = doc_data.read()
-                    doc = doc_info.query.filter(and_(doc_info.order_refnum == order.ref_num,
-                                             doc_info.order_refyear == order.ref_year)).first()
-                    if doc != None: 
-                        doc_entry = doc.update(
-                        doc_data=doc_content
-                        )
-                    else:
+                    if doc_data != None :
+                        doc_content = doc_data.read()
                         doc_entry = doc_info(
-                        order_refnum = order.ref_num,
-                        order_refyear = order.ref_year,
-                        filename= str(order.ref_num)+"/"+str(order.ref_year),
-                        doc_data=doc_content)
-                        db.session.add(doc_entry)
+                        order_refnum = order_entry.ref_num,
+                        order_refyear = order_entry.ref_year,
+                        filename= str(validated_dict['ref_num'])+"/"+str(validated_dict['ref_year']),
+                        doc_data=doc_content
+                    )
+                        db.session.add(doc_entry)   
+                    
+                else:
+                    app.logger.debug("update")
+
+                    order_entry = order.update(
+                    subject=validated_dict['subject'],
+                    doc_date=validated_dict['doc_date'],
+                    ref_name=name_list,
+                    user_id=validated_dict['user_id']
+                    )
+                    if doc_data != None :
+                        doc_content = doc_data.read()
+                        doc = doc_info.query.filter(and_(doc_info.order_refnum == order.ref_num,
+                                                doc_info.order_refyear == order.ref_year)).first()
+                        if doc != None: 
+                            doc_entry = doc.update(
+                            doc_data=doc_content
+                            )
+                        else:
+                            doc_entry = doc_info(
+                            order_refnum = order.ref_num,
+                            order_refyear = order.ref_year,
+                            filename= str(order.ref_num)+"/"+str(order.ref_year),
+                            doc_data=doc_content)
+                            db.session.add(doc_entry)
+                    
+
+                db.session.commit()
                 
-
-            db.session.commit()
-            
             return home()
-
+    else:
+        flash("You do not have permission.")
         return home()
     return render_template("project/index_table.html")
 
@@ -282,27 +289,34 @@ def remove():
         app.logger.debug(result)
         num = result.get('ref_num', '')
         year = result.get('ref_year', '')
-        try:
-            #contact = Contact.query.get(id_)
-            order = order_info.query.filter(and_(order_info.ref_num == int(num) ,
-                                             order_info.ref_year == int(year))).first()
-            app.logger.debug("order :",order)
-            latest_order = order_info.query.order_by(order_info.ref_num.desc(),order_info.ref_year.desc()).first()
-            doc = doc_info.query.filter(and_(doc_info.order_refnum == order.ref_num,
-                                             doc_info.order_refyear == order.ref_year)).first()
-            app.logger.debug("latest :",latest_order)
-            if doc != None:
-                db.session.delete(doc)
-            if order != latest_order:
-                db.session.delete(order)
-            else :
-                order.mini_update(
-                                    subject=None
-                                  )
-            db.session.commit()
-        except Exception as ex:
-            app.logger.debug(ex)
-            raise
+        access_token = session['access_token']
+        user_data = get_user_data(access_token)
+        email = user_data.get('cmuitaccount')
+        user = User.query.filter_by(email=email).first()
+        if user.role == True:
+            try:
+                #contact = Contact.query.get(id_)
+                order = order_info.query.filter(and_(order_info.ref_num == int(num) ,
+                                                order_info.ref_year == int(year))).first()
+                app.logger.debug("order :",order)
+                latest_order = order_info.query.order_by(order_info.ref_num.desc(),order_info.ref_year.desc()).first()
+                doc = doc_info.query.filter(and_(doc_info.order_refnum == order.ref_num,
+                                                doc_info.order_refyear == order.ref_year)).first()
+                app.logger.debug("latest :",latest_order)
+                if doc != None:
+                    db.session.delete(doc)
+                if order != latest_order:
+                    db.session.delete(order)
+                else :
+                    order.mini_update(
+                                        subject=None
+                                    )
+                db.session.commit()
+            except Exception as ex:
+                app.logger.debug(ex)
+                raise
+        else:
+            flash("You do not have permission to delete.")
     return home()
 
 @app.route('/search')
@@ -311,8 +325,12 @@ def search():
         # User is already authenticated, retrieve user data
         access_token = session['access_token']
         user_data = get_user_data(access_token)
+        email = user_data.get('cmuitaccount')
+        user = User.query.filter_by(email=email).first()
         if user_data:
             user_data_ = {
+                'id' : user.id,
+                'role' : user.role,
                 'name': user_data.get('firstname_TH') + " " + user_data.get('lastname_TH'),
                 'email': user_data.get('cmuitaccount')
                 }
@@ -332,8 +350,12 @@ def access():
     if 'access_token' in session:
         access_token = session['access_token']
         user_data = get_user_data(access_token)
+        email = user_data.get('cmuitaccount')
+        user = User.query.filter_by(email=email).first()
         if user_data:
             user_data_ = {
+                'id' : user.id,
+                'role' : user.role,
                 'name': user_data.get('firstname_TH') + " " + user_data.get('lastname_TH'),
                 'email': user_data.get('cmuitaccount')
             }
@@ -365,66 +387,74 @@ def user_data():
 
 @app.route('/user_form' , methods=('GET', 'POST'))
 def user_form():
-    if request.method == 'POST':
-        app.logger.debug("posted activate")
-        validated = True
-        validated_dict = dict()
-        valid_keys = ['role','email']
+    access_token = session['access_token']
+    user_data = get_user_data(access_token)
+    email = user_data.get('cmuitaccount')
+    user = User.query.filter_by(email=email).first()
+    if user.role == True:
+        if request.method == 'POST':
+            app.logger.debug("posted activate")
+            validated = True
+            validated_dict = dict()
+            valid_keys = ['role','email']
 
-        # Access the uploaded file using request.files
-        id_ = request.form.get('id','')
-        # validate the input
-        for key in request.form:
-            app.logger.debug(key)
-            if key not in valid_keys:
-                continue
-            value = request.form[key].strip()
-            if not value or value == 'undefined':
-                validated = False
-                break
-            app.logger.debug(value)
-            validated_dict[key] = value
+            # Access the uploaded file using request.files
+            id_ = request.form.get('id','')
+            # validate the input
+            for key in request.form:
+                app.logger.debug(key)
+                if key not in valid_keys:
+                    continue
+                value = request.form[key].strip()
+                if not value or value == 'undefined':
+                    validated = False
+                    break
+                app.logger.debug(value)
+                validated_dict[key] = value
 
-        if validated:
-            if not id_:
-                app.logger.debug("add new user")
-                app.logger.debug("role1 : " + validated_dict['role'])
-                # Create a new Document object with the uploaded file
-                if validated_dict['role'] == "True":
-                    role_ = True
-                elif validated_dict['role'] == "False":
-                    role_ = False
-                app.logger.debug("role2 : " + str(role_))
-                user_entry = User(
-                firstname="",
-                lastname="",
-                role=role_,
-                email=validated_dict['email']
-                )
-                db.session.add(user_entry)
-            else:
-                user = User.query.get(id_)
-                app.logger.debug("role : " + validated_dict['role'])
-                if validated_dict['role'] == "True":
-                    role_ = True
-                elif validated_dict['role'] == "False":
-                    role_ = False
-                app.logger.debug("role2 : " + str(role_))
-                user_entry = user.update(
-                email=validated_dict['email'],
-                role=role_
-                )
-            #     if doc_data != None :
-            #         doc_content = doc_data.read()
-            #         doc = doc_info.query.filter(doc_info.order_id == id_).first()   
-            #         doc_entry = doc.update(
-            #         filename= str(validated_dict['ref_num'])+"/"+str(validated_dict['ref_year']),
-            #         doc_data=doc_content
-            #         )
-            db.session.commit()
+            if validated:
+                if not id_:
+                    app.logger.debug("add new user")
+                    app.logger.debug("role1 : " + validated_dict['role'])
+                    # Create a new Document object with the uploaded file
+                    if validated_dict['role'] == "True":
+                        role_ = True
+                    elif validated_dict['role'] == "False":
+                        role_ = False
+                    app.logger.debug("role2 : " + str(role_))
+                    user_entry = User(
+                    firstname="",
+                    lastname="",
+                    role=role_,
+                    email=validated_dict['email']
+                    )
+                    db.session.add(user_entry)
+                else:
+                    user = User.query.get(id_)
+                    app.logger.debug("role : " + validated_dict['role'])
+                    if validated_dict['role'] == "True":
+                        role_ = True
+                    elif validated_dict['role'] == "False":
+                        role_ = False
+                    app.logger.debug("role2 : " + str(role_))
+                    user_entry = user.update(
+                    email=validated_dict['email'],
+                    role=role_
+                    )
+                #     if doc_data != None :
+                #         doc_content = doc_data.read()
+                #         doc = doc_info.query.filter(doc_info.order_id == id_).first()   
+                #         doc_entry = doc.update(
+                #         filename= str(validated_dict['ref_num'])+"/"+str(validated_dict['ref_year']),
+                #         doc_data=doc_content
+                #         )
+                db.session.commit()
+                return home()
+
             return home()
-
-        return home()  
+        else:
+            flash("You do not have permission to delete.")
+            return home()
     return ''
 
 @app.route("/data")
@@ -492,14 +522,21 @@ def user_remove():
         result = request.form.to_dict()
         app.logger.debug(result)
         id_ = result.get('id', '')
-        try:
-            #contact = Contact.query.get(id_)
-            order = User.query.get(id_)
-            db.session.delete(order)
-            db.session.commit()
-        except Exception as ex:
-            app.logger.debug(ex)
-            raise
+        access_token = session['access_token']
+        user_data = get_user_data(access_token)
+        email = user_data.get('cmuitaccount')
+        user = User.query.filter_by(email=email).first()
+        if user.role == True:
+            try:
+                #contact = Contact.query.get(id_)
+                order = User.query.get(id_)
+                db.session.delete(order)
+                db.session.commit()
+            except Exception as ex:
+                app.logger.debug(ex)
+                raise
+        else:
+            flash("You do not have permission to delete.")
     return home()
 
 # @app.route('/login', methods=('GET', 'POST'))
